feat(linter/typescript): implement prefer-destructuring rule
