+++ crates/oxc_linter/src/rules.rs
@@ -311,6 +311,7 @@ pub(crate) mod typescript {
     pub mod only_throw_error;
     pub mod parameter_properties;
     pub mod prefer_as_const;
+    pub mod prefer_destructuring;
     pub mod prefer_enum_initializers;
     pub mod prefer_find;
     pub mod prefer_for_of;
