+++ crates/oxc_linter/src/rules/shared/eslint_typescript/mod.rs
@@ -1,0 +1,1 @@
+pub mod prefer_destructuring;
