+++ crates/oxc_linter/src/rules/shared/mod.rs
@@ -1,5 +1,7 @@
+mod eslint_typescript;
 mod eslint_unicorn;
 mod jest_vitest;
 
+pub use eslint_typescript::*;
 pub use eslint_unicorn::*;
 pub use jest_vitest::*;
